# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  seedservice.PartyAdminsData:
    properties:
      id:
        description: partyID
        type: integer
      party_admin:
        items:
          type: integer
        type: array
      super_party_admin:
        items:
          type: integer
        type: array
    type: object
  seedservice.SeedAdminsRequest:
    properties:
      admins:
        items:
          type: integer
        type: array
      parties:
        items:
          additionalProperties:
            $ref: '#/definitions/seedservice.PartyAdminsData'
          type: object
        type: array
    type: object
  seedservice.SeedUserRequest:
    properties:
      account_status:
        type: string
      avatar:
        type: string
      current_city:
        type: integer
      current_country:
        type: integer
      current_lga:
        type: integer
      current_state:
        type: integer
      date_of_birth:
        type: string
      education_level:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      home_address:
        type: string
      is_politician:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      marital_status:
        type: string
      middle_name:
        type: string
      num:
        type: integer
      occupation_id:
        type: integer
      party_id:
        type: integer
      password:
        type: string
      phone:
        type: string
      religion:
        type: string
      state_of_origin:
        type: integer
      username:
        type: string
      verification_type_id:
        type: integer
    type: object
  seedservice.SimulateElectionResultsRequest:
    properties:
      limit:
        description: max PUs to populate (0 = all)
        type: integer
      max_votes_per_pu:
        description: optional max votes (default 750)
        type: integer
      min_votes_per_pu:
        description: optional min votes (default 100)
        type: integer
      run_full_rollup:
        description: immediately run sequential rollup queries (default false)
        type: boolean
      trigger_realtime:
        description: dispatch Asynq tasks (default true)
        type: boolean
    type: object
...
host: localhost:4100
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for the Free9ja platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Free9ja API
  version: "1.0"
paths:
...
  /seed/admins:
    post:
      consumes:
      - application/json
      description: Batch registers roles for admins from formatted JSON data
      parameters:
      - description: List of admins to seed
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/seedservice.SeedAdminsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Admins seeded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to seed admins
          schema:
            additionalProperties: true
            type: object
      summary: Seed admins and party admins
      tags:
      - Seed
  /seed/elections/{id}/simulate-results:
    post:
      consumes:
      - application/json
      description: Populates mock consensus final results for all eligible polling
        units of an election and triggers rollups
      parameters:
      - description: Election ID
        in: path
        name: id
        required: true
        type: integer
      - description: Simulation options
        in: body
        name: request
        schema:
          $ref: '#/definitions/seedservice.SimulateElectionResultsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Results simulated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to simulate election results
          schema:
            additionalProperties: true
            type: object
      summary: Simulate election polling unit results
      tags:
      - Seed
  /seed/users:
    post:
      consumes:
      - application/json
      description: Batch registers testing users from formatted JSON data
      parameters:
      - description: List of users to seed
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/seedservice.SeedUserRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Users seeded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to seed users
          schema:
            additionalProperties: true
            type: object
      summary: Seed testing users
      tags:
      - Seed
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your access token.
    in: header
    name: Authorization
    type: apiKey
  RefreshToken:
    description: Type your refresh token.
    in: header
    name: RefreshToken
    type: apiKey
swagger: "2.0"